jsonrpc - make the transport protocol work correctly
Jolie already had this (broken) jsonrpc implementation - but directly
over sockets. After long investigations I found out the reason of my
communication problems: Jolie uses json-simple (http://code.google.com/p/json-simple/)
as its JSON parser. This one has the attitude that it always parses
until EOF on an InputStream is reached (and not until the JSON
expression is completed). This works well for files and streams on
restricted strings, but is absolutely problematic on sockets. Moreover
the author knows about this issue and has no intention to fix it:
http://code.google.com/p/json-simple/issues/detail?id=66

Hence the solution is to fall back to some sort of transport protocol
with content length information which allows us to generate bounded
streams. I chose HTTP since it is pretty standard for jsonrpc and widely
used in Jolie, and made the necessary changes.