HttpProtocol - fix channel closing
By https://tools.ietf.org/html/rfc7230#section-6.6 an HTTP server should not be closing
a channel immediately after a request before delivering the response.

So for now keep this close mechanism only for the HTTP client. For the server we would
need to save save this directive somewhere and apply it only *after* delivering the whole
response.


Former-commit-id: eb4aa9ed473fbc6c5fe6d02749e963b3a032f5c8